pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nanoplm"
version = "0.1.0"
description = "Make your own protein language model"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Peyman Vahidi", email = "peyman.vahidi@tum.de"},
]
dependencies = [
    "accelerate>=1.7.0",
    "biopython>=1.85",
    "click>=8.1.8",
    "dvc>=3.59.1",
    "h5py>=3.13.0",
    "onnx>=1.17.0",
    "pyyaml>=6.0.2",
    "safetensors>=0.4.0",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.0",
    "torch>=2.5.4",
    "tqdm>=4.67.1",
    "transformers>=4.50.2",
    "wandb>=0.19.11",
    "certifi>=2024.7.4",
    "ipython>=8.3.0",
]

[project.scripts]
nanoplm = "nanoplm.cli:cli"
nanoplm-data = "nanoplm.cli.data:data"
nanoplm-train = "nanoplm.cli.train:train"

[tool.setuptools.package-data]
"nanoplm.models.student" = ["*.csv"]
